✨ We have a live graph
- gotta get candidate names
- gotta pretty it up
- gotta get better test data
- gotta make sure it's live
