+++ src/components/WaitForVoters/WaitForVoters.tsx
@@ -3,7 +3,7 @@ import "./waitForVoters.css"
 import { runTransaction } from "atom.io"
 import { useO } from "atom.io/react"
 
-import BarChart from "~/src/components/BarChart/BarChart2"
+import BarChart from "~/src/components/BarChart/BarChart"
 import { joinElectionTX, retractSubmittedBallot } from "~/src/lib/election"
 import { candidatesInCurrentElectionSelector } from "~/src/lib/election-candidates"
 import {
@@ -11,7 +11,7 @@ import {
 	currentElectionVotersSelector,
 } from "~/src/lib/election-voters"
 
-import type { ElectionState } from "../../types"
+import type { ElectionState, GraphableCandidateVote } from "../../types"
 
 type WaitForVotersProps = {
 	targetState: ElectionState
@@ -45,11 +45,27 @@ function getLiveCandidateVotes(currentElectionVoters: CurrentElectionVoters[]): 
 	return output
 }
 
+function transformToGraphable(liveCandidateVotes: LiveCandidateVotes): GraphableCandidateVote[] {
+	const output: GraphableCandidateVote[] = Object.entries(liveCandidateVotes).map(
+		([candidateId, tierTotals]) => ({
+			id: candidateId,
+			t1: tierTotals[0],
+			t2: tierTotals[1],
+			t3: tierTotals[2],
+			t4: tierTotals[3],
+			t5: tierTotals[4],
+			t6: tierTotals[5],
+		}),
+	)
+	return output
+}
+
 function WaitForVoters({ targetState }: WaitForVotersProps): JSX.Element {
 	const currentElectionVoters = useO(currentElectionVotersSelector)
 	const joinElection = runTransaction(joinElectionTX)
 
 	const liveCandidateVotes = getLiveCandidateVotes(currentElectionVoters)
+	const graphableCandidateVotes = transformToGraphable(liveCandidateVotes)
 	console.log(JSON.stringify(liveCandidateVotes, null, 2))
 	return (
 		<div className="waitForVoters">
@@ -58,7 +74,12 @@ function WaitForVoters({ targetState }: WaitForVotersProps): JSX.Element {
 					<h1>Waiting for voters...</h1>
 					<h2 style={{ padding: `10px` }}>Current voters:</h2>
 					<ul>
-						<BarChart width={500} height={300} />
+						<BarChart
+							data={graphableCandidateVotes}
+							width={500}
+							height={300}
+							margin={{ top: 20, left: 50, right: 40, bottom: 100 }}
+						/>
 					</ul>
 				</div>
 			) : (
